fix:migrationファイルの修正

+++ database/migrations/2022_10_08_062233_create_comments_table.php
@@ -12,7 +12,7 @@ class CreateCommentsTable extends Migration
             $table->increments('id');
             $table->string('title', 30);
             $table->string('body', 300);
-            $table->integer('rating');
+            $table->tinyInteger('rating');
             $table->timestamps();
             $table->softDeletes();
             $table->integer('snack_id')->unsigned();
